pyproject.toml
[tool.poetry]
name = "koyeb-nb2"
version = "0.1.0"
description = "koyeb-nb2"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
# python = "^3.7"
python = "^3.7.3"
nonebot2 = "^2.0.0-alpha.13"
nonebot-plugin-guess = "^0.1.5"
chardet = "^4.0.0"
Jinja2 = "^3.0.1"
nonebot-adapter-cqhttp = "^2.0.0-alpha.13"
arrow = "^1.1.1"
aiocqhttp = "^1.4.1"
readability-lxml = "^0.8.1"
html2text = "^2020.1.16"
jsonpath-ng = "^1.5.3"
nonebot-plugin-apscheduler = "^0.1.2"
pyquery = "^1.4.3"
joblib = "^1.0.1"
pdir2 = "^0.3.2"
httpx = "0.17.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"
poethepoet = "^0.10.0"
respx = "~=0.16.0"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"

[nonebot.plugins]
plugins = []
plugin_dirs = ["koyeb_nb2/plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
